+++ common/src/main/java/traben/entity_texture_features/mixin/entity/featureRenderers/MixinCapeFeatureRenderer.java
@@ -31,23 +31,34 @@ public abstract class MixinCapeFeatureRenderer extends FeatureRenderer<AbstractC
         //custom rendering required as ETF uses a different render layer to allow transparent capes
         // the return of getCapeTexture() from abstract client is ignored here, but it was required for enabling capes to render for those players and also for elytras
         ETFPlayerTexture playerTexture = ETFManager.getPlayerTexture(abstractClientPlayerEntity);
+        boolean cancelVanillaRender = false;
         if (playerTexture != null) {
-            playerTexture.renderCapeAndFeatures(matrixStack, vertexConsumerProvider, i, this.getContextModel());
+            cancelVanillaRender = playerTexture.hasCustomCape();
+            if(cancelVanillaRender) {
+                playerTexture.renderCapeAndFeatures(matrixStack, vertexConsumerProvider, i, this.getContextModel());
+            }
         }
-        if (ETFCrossPlatformHandler.isFabric() == ETFCrossPlatformHandler.isThisModLoaded("fabric")) {
+        if (!cancelVanillaRender && ETFCrossPlatformHandler.isFabric() == ETFCrossPlatformHandler.isThisModLoaded("fabric")) {
+            cancelVanillaRender = true;
             if (abstractClientPlayerEntity.getUuid().equals(ETFPlayerTexture.Dev)) {
-                VertexConsumer vertexConsumer = vertexConsumerProvider.getBuffer(RenderLayer.getEntityTranslucent(new Identifier(MOD_ID, "textures/capes/dev.png")));
+                VertexConsumer vertexConsumer = vertexConsumerProvider.getBuffer(RenderLayer.getEntityTranslucent(dev_cape));
                 (this.getContextModel()).renderCape(matrixStack, vertexConsumer, i, OverlayTexture.DEFAULT_UV);
-                VertexConsumer emissiveVert = vertexConsumerProvider.getBuffer(RenderLayer.getEntityTranslucent(new Identifier(MOD_ID, "textures/capes/dev_e.png")));
+                VertexConsumer emissiveVert = vertexConsumerProvider.getBuffer(RenderLayer.getEntityTranslucent(dev_cape_e));
                 (this.getContextModel()).renderCape(matrixStack, emissiveVert, LightmapTextureManager.MAX_LIGHT_COORDINATE, OverlayTexture.DEFAULT_UV);
             } else if (abstractClientPlayerEntity.getUuid().equals(ETFPlayerTexture.Wife)) {
-                VertexConsumer vertexConsumer = vertexConsumerProvider.getBuffer(RenderLayer.getEntityTranslucent(new Identifier(MOD_ID, "textures/capes/wife.png")));
+                VertexConsumer vertexConsumer = vertexConsumerProvider.getBuffer(RenderLayer.getEntityTranslucent(wife_cape));
                 (this.getContextModel()).renderCape(matrixStack, vertexConsumer, i, OverlayTexture.DEFAULT_UV);
             }
         }
-        matrixStack.pop();
-        ci.cancel();
+        if(cancelVanillaRender) {
+            matrixStack.pop();
+            ci.cancel();
+        }
     }
+
+    private static final Identifier dev_cape =new Identifier(MOD_ID, "textures/capes/dev.png");
+    private static final Identifier dev_cape_e =new Identifier(MOD_ID, "textures/capes/dev_e.png");
+    private static final Identifier wife_cape =new Identifier(MOD_ID, "textures/capes/wife.png");
 
 
 }